requirements.txt
fastapi==0.111.1
pydantic==1.10.2
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2
fastapi-pagination==0.12.26
SQLAlchemy==2.0.31
asyncpg==0.29.0
multimethod==1.12
psycopg2-binary==2.9.3
environs==11.0.0
greenlet==3.0.3
